+++ libraries/rotaryDecoder/rotaryDecoder.h
@@ -2,16 +2,16 @@
 //
 //    FILE: rotaryDecoder.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.1
+// VERSION: 0.3.0
 //    DATE: 2021-05-08
-// PURPOSE: rotary decoder library for Arduino
+// PURPOSE: Arduino library for rotary decoder
 //     URL: https://github.com/RobTillaart/rotaryDecoder
 
 
 #include "Arduino.h"
 #include "Wire.h"
 
-#define ROTARY_DECODER_LIB_VERSION         (F("0.2.1"))
+#define ROTARY_DECODER_LIB_VERSION         (F("0.3.0"))
 
 
 class rotaryDecoder
@@ -37,17 +37,24 @@ public:
   int32_t getValue(uint8_t re);
   void    setValue(uint8_t re, int32_t value = 0);
 
+
+  //  READ - WRITE interface
+  uint8_t read1(uint8_t pin);
+  bool    write1(uint8_t pin, uint8_t value);
+  uint8_t read8();
+  bool    write8(uint8_t value);
+
+
   //  DEBUG
   uint8_t getLastPosition(uint8_t re);
 
 
-private:
+protected:
   uint8_t   _count = 0;
   uint8_t   _lastValue = 0;
   uint8_t   _lastPos[4] = { 0, 0, 0, 0 };
   int32_t   _encoder[4] = { 0, 0, 0, 0 };
 
-  uint8_t   _read8();
   uint8_t   _address;
   TwoWire * _wire;
 };