+++ libraries/rotaryDecoder/CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.3.0] - 2024-02-14
+- Fix #10
+- add **read1(pin)**
+- add **write1(pin, value)** experimental see #10
+- made **read8()** and **write8()** public for faster multi pin access
+- update readme.md
+
+----
+
 ## [0.2.1] - 2024-01-06
 - Fix URL in examples
 - minor edits